+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
-
+- :book: BREAKING CHANGE: The CLI interface was refactored to support subcommands. To build a bookmarks app, use the `build` command: `static-marks [options] build <files>`.
+- :boom: BREAKING CHANGE: Input file argument(s) are now required (previously the current folder was used as fallback).
 
 ## 1.0.8 (2019-04-28)
 - :sparkles: Add support for custom web page templates.